rename `_types` to be public
