+++ pydantic_core/__init__.py
@@ -1,4 +1,4 @@
 from ._pydantic_core import PydanticValueError, SchemaError, SchemaValidator, ValidationError, __version__
-from ._types import Config, Schema
+from .schema_types import Config, Schema
 
 __all__ = '__version__', 'Config', 'Schema', 'SchemaValidator', 'SchemaError', 'ValidationError', 'PydanticValueError'
