+++ pydantic_core/schema_types.py
